requirements.txt
customtkinter
Pillow
numpy<2
matplotlib
pyserial
cartopy
roboticstoolbox-python

